requirements.txt
py-postgresql>=1.1.0
git+https://github.com/Toblerity/Shapely.git
